Cleaning and training flower classification
In this first two files we can find two different parts of a computer vision problem. The classification is made with only two types of flowers, sunflowers and roses as an approach to demonstrate the capabilities of the methods which have been used. An unsupervised innovative approach for the classification of both has been used.